+++ src/Array.cpp
@@ -615,6 +615,7 @@ DEFINE_ARRAY_FUNC1(return,resize);
 
 DEFINE_ARRAY_FUNC1(return,concat);
 DEFINE_ARRAY_FUNC0(return,iterator);
+DEFINE_ARRAY_FUNC0(return,keyValueIterator);
 DEFINE_ARRAY_FUNC1(return,join);
 DEFINE_ARRAY_FUNC0(return,pop);
 DEFINE_ARRAY_FUNC0(return,copy);
@@ -644,6 +645,7 @@ hx::Val ArrayBase::__Field(const String &inString, hx::PropertyAccess inCallProp
    if (inString==HX_CSTRING("insert")) return insert_dyn();
    if (inString==HX_CSTRING("copy")) return copy_dyn();
    if (inString==HX_CSTRING("iterator")) return iterator_dyn();
+   if (inString==HX_CSTRING("keyValueIterator")) return keyValueIterator_dyn();
    if (inString==HX_CSTRING("join")) return join_dyn();
    if (inString==HX_CSTRING("pop")) return pop_dyn();
    if (inString==HX_CSTRING("push")) return push_dyn();
@@ -681,6 +683,7 @@ static String sArrayFields[] = {
    HX_CSTRING("concat"),
    HX_CSTRING("insert"),
    HX_CSTRING("iterator"),
+   HX_CSTRING("keyValueIterator"),
    HX_CSTRING("join"),
    HX_CSTRING("copy"),
    HX_CSTRING("pop"),
@@ -804,6 +807,7 @@ Dynamic VirtualArray_obj::func##_dyn()  { return new VirtualArray_##func(this); 
 DEFINE_VARRAY_FUNC1(return,concat);
 DEFINE_VARRAY_FUNC2(,insert);
 DEFINE_VARRAY_FUNC0(return,iterator);
+DEFINE_VARRAY_FUNC0(return,keyValueIterator);
 DEFINE_VARRAY_FUNC1(return,join);
 DEFINE_VARRAY_FUNC0(return,pop);
 DEFINE_VARRAY_FUNC0(return,copy);
@@ -877,6 +881,7 @@ hx::Val VirtualArray_obj::__Field(const String &inString, hx::PropertyAccess inC
    if (inString==HX_CSTRING("insert")) return insert_dyn();
    if (inString==HX_CSTRING("copy")) return copy_dyn();
    if (inString==HX_CSTRING("iterator")) return iterator_dyn();
+   if (inString==HX_CSTRING("keyValueIterator")) return keyValueIterator_dyn();
    if (inString==HX_CSTRING("join")) return join_dyn();
    if (inString==HX_CSTRING("pop")) return pop_dyn();
    if (inString==HX_CSTRING("push")) return push_dyn();